refactor: Added Common Directory
Separated:
1.  Abstract models
2. Common middlewares
3. Common Mixins

into the common directory
